Delegate SecDiv creation and deletion to kernel via kernel API
To have a consistent range table, kernel should be responsible for
creating or deleting (i.e., blanking permissions of) security
divisions. This is now delegated to the kernel via the API. This
change is in line with the RISC-V instruction implementation that
forces the CPU to fault when instructions on a SecDiv that hasn't
been explicitly created beforehand are called.