Refactor code into easier to understand functions
The refactoring also includes also only using two SecDivs, the
original rootserver and the client thread. The model with three
threads before was not flexible enough and doesn't have any
influence on the benchmarks. Since we also only have two threads/
processes in the classic IPC benchmarks, this makes the code more
comparable and readable.