Increase static memory pool for allocator
With the smaller pool, the allocator all of a sudden failed. Since
the error messages were unfortunately not very helpful, the deeper
root for the issue remains unknown.
This patch is therefore just a workaround (which could also be a
proper solution but this is not evident to us yet).