Remove forcing no kernel optimization
The kernel optimization was previously necessary due to it
triggering compiler errors if the kernel was built with
optimizations enabled. We could now get around those errors.
Kernel optimization is not important for our case since the
performance of SecCells is due to its userspace focus.
Nevertheless, this brings building the kernel on par with the page-
based benchmarks.