Change SCCount semantics to SCExcl
Instead of having an SCCount instruction that returns the number
of SecDivs having access to a specified location, we only require
an SCExcl instruction returning a boolean flag for exclusivity.
The changes in the test code as well as libseccells reflect this
change in semantics.

In addition, the debugging script accounts for an updated kernel
entrypoint.